refactor: Circular ProgressBar
- 메인 회색원과 gradient progressBar분리
- 주석추가
